requirements.txt
asgiref==3.5.2
Django==4.1
Pillow==9.2.0
sqlparse==0.4.2
django-simple-captcha==0.5.17